Improve the API to save, load and clear account data using the keystore
The implementation will save the account data when user signs in to avoid
the need of constantly sign in after the app is closed.

When the user opens the app, the first task that must be done is to verify
if there is an existing saved account to restore the session if so. This task,
must be done, if possible, in a splash screen.

When the user signs out, the app clear all the account data from the
keystore.

IMPORTANT NOTE: If any of this tasks (save, load or clear) cannot be done,
the app should continue working as if it wasn't ever be called.